# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Class:
    properties:
      academic_year:
        type: string
      created_at:
        type: string
      homeroom:
        type: string
      id:
        type: integer
      level:
        type: string
      name:
        type: string
      school:
        $ref: '#/definitions/models.School'
      school_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.CreateStudent:
    properties:
      class_id:
        type: integer
      email:
        type: string
      name:
        type: string
      school_id:
        type: integer
      type:
        enum:
        - junior
        - senior
        example: junior
        type: string
    required:
    - email
    - name
    - type
    type: object
...
  models.LoginUser:
    properties:
      password:
        type: string
      school_id:
        type: integer
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.School:
    properties:
      address:
        type: string
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Student:
    properties:
      class:
        $ref: '#/definitions/models.Class'
      class_id:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      school:
        $ref: '#/definitions/models.School'
      school_id:
        type: integer
      type:
        enum:
        - junior
        - senior
        example: junior
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateStudent:
    properties:
      class_id:
        type: integer
      email:
        type: string
      name:
        type: string
      school_id:
        type: integer
      type:
        enum:
        - junior
        - senior
        example: senior
        type: string
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-Raport API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user using username and password, returns a JWT
        token if successful
      parameters:
      - description: User login object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: JWT Token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Authenticate a user
      tags:
      - user
  /register:
    post:
      consumes:
      - application/json
      description: Registers a new user with the given username and password
      parameters:
      - description: User registration object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully registered
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Register a new user
      tags:
      - user
  /students:
    get:
      description: Get a list of all students with optional pagination
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - default: 10
        description: Limit for pagination
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of students
          schema:
            items:
              $ref: '#/definitions/models.Student'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all students with pagination
      tags:
      - students
    post:
      consumes:
      - application/json
      description: Create a new student with kindergarten type (junior/senior)
      parameters:
      - description: Create student object
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateStudent'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created student
          schema:
            $ref: '#/definitions/models.Student'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Create a new student
      tags:
      - students
  /students/{id}:
    delete:
      description: Delete student by ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted student
          schema:
            type: string
        "404":
          description: Student not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a student by ID
      tags:
      - students
    get:
      description: Get details of a student by ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved student
          schema:
            $ref: '#/definitions/models.Student'
        "404":
          description: Student not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Find a student by ID
      tags:
      - students
    put:
      consumes:
      - application/json
      description: Update student details and kindergarten type (junior/senior)
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      - description: Update student object
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStudent'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated student
          schema:
            $ref: '#/definitions/models.Student'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Student not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update a student by ID
      tags:
      - students
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  JwtAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"